ENT-1906: Publish DJVM artifact with its dependency information. Also migrate enclavelet code into testing:common jar.
