+++ djvm/build.gradle
@@ -11,10 +11,14 @@ ext {
     asm_version = '6.1.1'
 }
 
+configurations {
+    testCompile.extendsFrom shadow
+}
+
 dependencies {
-    compile "org.jetbrains.kotlin:kotlin-stdlib-jdk8:$kotlin_version"
-    compile "org.jetbrains.kotlin:kotlin-reflect:$kotlin_version"
-    compile "org.slf4j:slf4j-api:$slf4j_version"
+    shadow "org.jetbrains.kotlin:kotlin-stdlib-jdk8:$kotlin_version"
+    shadow "org.jetbrains.kotlin:kotlin-reflect:$kotlin_version"
+    shadow "org.slf4j:slf4j-api:$slf4j_version"
 
     // ASM: byte code manipulation library
     compile "org.ow2.asm:asm:$asm_version"
@@ -22,7 +26,7 @@ dependencies {
     compile "org.ow2.asm:asm-commons:$asm_version"
 
     // Classpath scanner
-    compile "io.github.lukehutch:fast-classpath-scanner:$fast_classpath_scanner_version"
+    shadow "io.github.lukehutch:fast-classpath-scanner:$fast_classpath_scanner_version"
 
     // Test utilities
     testCompile "junit:junit:$junit_version"
@@ -35,11 +39,6 @@ jar.enabled = false
 shadowJar {
     baseName 'corda-djvm'
     classifier ''
-    dependencies {
-        exclude(dependency('org.jetbrains.*:.*:.*'))
-        exclude(dependency('org.slf4j:.*:.*'))
-        exclude(dependency('io.github.lukehutch:.*:.*'))
-    }
     relocate 'org.objectweb.asm', 'djvm.org.objectweb.asm'
 }
 assemble.dependsOn shadowJar
@@ -49,6 +48,7 @@ artifacts {
 }
 
 publish {
+    dependenciesFrom configurations.shadow
     disableDefaultJar true
     name shadowJar.baseName
 }
